proyecto_1_hpt_img/requirements.txt
numpy==1.24.4
pandas==2.0.3
pickle5==0.0.11
scikit-learn==1.3.2
xgboost==2.1.4
gcsfs==2023.6.0
cloudml-hypertune==0.1.0.dev6